+++ src/gitversioniser/config/commit_change_tags/__init__.py
@@ -1,0 +1,5 @@
+from .abstract import CommitChangeTags
+from .core.classic_changelog import ClassicChangelog
+
+
+__all__ = ['CommitChangeTags', 'ClassicChangelog']