+++ src/gitversioniser/config/commit_change_tags/commit_tag.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from typing import Optional
 
 from gitversioniser.helpers.changelog_category import ChangelogCategory
 
@@ -7,6 +8,7 @@ from gitversioniser.helpers.changelog_category import ChangelogCategory
 class CommitTag:
     pattern: str
     changelog_category: ChangelogCategory
+    emoji_representation: Optional[str] = None
 
     def __str__(self):
         return self.pattern