+++ src/gitversioniser/config/commit_change_tags/core/__init__.py
@@ -1,0 +1,7 @@
+from .angular_commits import AngularCommits
+from .classic_changelog import ClassicChangelog
+from .classic_changelog_extended import ClassicChangelogExtended
+from .conventional_commits import ConventionalCommits
+
+
+__all__ = ['AngularCommits', 'ClassicChangelog', 'ClassicChangelogExtended', 'ConventionalCommits']