+++ src/gitversioniser/config/commit_changelog_tags/__init__.py
@@ -1,5 +1,0 @@
-from .abstract import CommitChangelogTags
-from .core.classic_changelog import ClassicChangelog
-
-
-__all__ = ['CommitChangelogTags', 'ClassicChangelog']