+++ src/gitversioniser/config/commit_changelog_tags/core/__init__.py
@@ -1,5 +1,0 @@
-from .angular_changelog import AngularChangelog
-from .classic_changelog import ClassicChangelog
-
-
-__all__ = ['AngularChangelog', 'ClassicChangelog']