+++ src/gitversioniser/domain/versioniser/routines/changelog/core/__init__.py
@@ -1,6 +1,6 @@
-from .commit_changelog_tags.commit_changelog_tags import CommitChangelogTags
+from .commit_change_tags.commit_change_tags import CommitChangeTags
 from .last_commit_message_as_description.last_commit_message_as_description import LastCommitMessageAsDescription
 from .null.null import Null
 
 
-__all__ = ['Null', 'LastCommitMessageAsDescription', 'CommitChangelogTags']
+__all__ = ['Null', 'LastCommitMessageAsDescription', 'CommitChangeTags']