+++ tests/v0/test_default_config_values.py
@@ -1,4 +1,4 @@
-from gitversioniser.config.commit_changelog_tags.commit_tag import CommitTag
+from gitversioniser.config.commit_change_tags.commit_tag import CommitTag
 from gitversioniser.helpers.changelog_category import ChangelogCategory
 from tests.v0.default_scenario_v0 import TestDefaultScenarioV0
 
@@ -14,7 +14,7 @@ class TestDefaultConfigValues(TestDefaultScenarioV0):
         self.assertEqual(self.config.routines.file_updater, 'VersioniseFiles')
         self.assertEqual(self.config.routines.commiting, 'PushOriginAmend')
         self.assertEqual(self.config.routines.tagging, 'Regular')
-        self.assertEqual(self.config.routines.changelog, 'CommitChangelogTags')
+        self.assertEqual(self.config.routines.changelog, 'CommitChangeTags')
 
     def test_default_configuration_credentials(self):
         self.assertEqual(self.config.credentials.username, "GitVersioniser")
@@ -37,9 +37,9 @@ class TestDefaultConfigValues(TestDefaultScenarioV0):
         self.assertEqual(self.config.patterns.increments.finalized.precedence, 9)
 
     def test_default_configuration_patterns_commit_tags(self):
-        self.assertEqual(self.config.patterns.commit_tags.added, CommitTag("A:", ChangelogCategory.ADDED))
-        self.assertEqual(self.config.patterns.commit_tags.changed, CommitTag("C:", ChangelogCategory.CHANGED))
-        self.assertEqual(self.config.patterns.commit_tags.deprecated, CommitTag("D:", ChangelogCategory.DEPRECATED))
-        self.assertEqual(self.config.patterns.commit_tags.removed, CommitTag("R:", ChangelogCategory.REMOVED))
-        self.assertEqual(self.config.patterns.commit_tags.security, CommitTag("S:", ChangelogCategory.SECURITY))
-        self.assertEqual(self.config.patterns.commit_tags.fixed, CommitTag("F:", ChangelogCategory.FIXED))
+        self.assertEqual(self.config.patterns.commit_tags.added, CommitTag("A:", ChangelogCategory.ADDED, "💠"))
+        self.assertEqual(self.config.patterns.commit_tags.fixed, CommitTag("F:", ChangelogCategory.FIXED, "🛠️"))
+        self.assertEqual(self.config.patterns.commit_tags.changed, CommitTag("C:", ChangelogCategory.CHANGED, "🔸"))
+        self.assertEqual(self.config.patterns.commit_tags.removed, CommitTag("R:", ChangelogCategory.REMOVED, "🗑️"))
+        self.assertEqual(self.config.patterns.commit_tags.security, CommitTag("S:", ChangelogCategory.SECURITY, "🔐"))
+        self.assertEqual(self.config.patterns.commit_tags.deprecated, CommitTag("D:", ChangelogCategory.DEPRECATED, "🔚"))